Move `build_dep_graph` into `rustc_incremental::persist::load`
Moving this code here makes the subsequent cleanup easier.
